backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai>=1.12.0
pydantic==2.5.0
pydantic-settings==2.1.0
sse-starlette==1.8.2
httpx>=0.25.0
httpcore>=1.0.0
supabase>=2.0.0

